refactor: Remove inheritance
After discussion with Patrick, as it is only one method, remove
the inheritance and let the derived class define its own method
from the two available
