+++ packages/cozy-stack-client/src/SharingCollection.js
@@ -1,4 +1,4 @@
-import DocumentCollection, { normalizeDoc } from './JsonApiDocumentCollection'
+import DocumentCollection, { normalizeDoc } from './DocumentCollection'
 import { isFile, isDirectory } from './FileCollection'
 import { uri } from './utils'
 
@@ -110,6 +110,8 @@ class SharingCollection extends DocumentCollection {
     )
   }
 }
+
+SharingCollection.normalizeDoctype = DocumentCollection.normalizeDoctypeJsonApi
 
 // Rules determine the behavior of the sharing when changes are made to the shared document
 // See https://github.com/cozy/cozy-stack/blob/master/docs/sharing-design.md#description-of-a-sharing
